Combine Cloud Vision samples into one file
